reset password

+++ src/utils/redis.ts
@@ -1,4 +1,6 @@
 import { createClient } from 'redis'
+import { generateOTP } from './nanoid'
+import { OTPReqBody } from '~/models/requests/users.requests'
 
 const client = createClient({
   username: 'default',
@@ -21,38 +23,43 @@ class RedisUtils {
     }
   }
 
-  async get(key: string) {
-    try {
-      const value = await client.get(key)
-      return value
-    } catch (error) {
-      throw error
-    }
+  async saveOTP({ email }: OTPReqBody) {
+    const otp = generateOTP()
+    await client.set(`${otp}:${email}`, otp, { EX: Number(process.env.REDIS_OTP_TOKEN_TTL) })
+    return otp
   }
 
-  async set(key: string, value: string) {
-    try {
-      return await client.set(key, value, { EX: 120 })
-    } catch (error) {
-      throw error
+  async verifyOTP(email: string, otp: string) {
+    const key = `${otp}:${email}`
+
+    // Kiểm tra OTP có tồn tại và đúng không
+    const storedOTP = await client.get(key)
+
+    if (!storedOTP) {
+      return false
     }
+
+    // OTP hợp lệ - xóa OTP sau khi đã sử dụng thành công
+    await client.del(key)
+    return true
   }
 
-  async del(key: string) {
-    try {
-      return await client.del(key)
-    } catch (error) {
-      throw error
-    }
+  async saveForgotPasswordToken(user_id: string, token: string) {
+    const key = `forgot_password:${user_id}`
+    await client.set(key, token, { EX: Number(process.env.REDIS_FORGOT_PASSWORD_TOKEN_TTL) })
   }
 
-  async exists(key: string) {
-    try {
-      const result = await client.exists(key)
-      return result
-    } catch (error) {
-      throw error
+  async verifyForgotPasswordToken(user_id: string, token: string) {
+    const key = `forgot_password:${user_id}`
+    const storedToken = await client.get(key)
+
+    if (!storedToken || storedToken !== token) {
+      return false
     }
+
+    // Token hợp lệ - xóa token sau khi đã sử dụng thành công
+    await client.del(key)
+    return true
   }
 }
